created new Interactive Map component
