+++ react-bricks/bricks/react-bricks-ui/index.ts
@@ -67,7 +67,7 @@ import Faq from './Faq/FaqItem'
 import Offices from './contacts/Offices/Offices'
 import Office from './contacts/Offices/Office'
 import ContactsForm from './contacts/ContactsForm/ContactsForm'
-import Map from './contacts/Map/Map'
+import Map from './contacts/Maps/Map'
 import FormBuilder from './contacts/FormBuilder/FormBuilder'
 import FormInput from './contacts/FormBuilder/FormInput'
 import FormTextarea from './contacts/FormBuilder/FormTextarea'
